pyproject.toml
[project]
name = "difflex"
version = "0.0.1"
description = "Smart file and directory comparison tool with 3-way diff, similarity detection, and external tool integration."
authors = [{ name = "kakehashi" }]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["diff", "directory", "file", "text", "image", "binary"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.urls]
"Homepage" = "https://github.com/kakehashi-inc/difflex"
"Bug Tracker" = "https://github.com/kakehashi-inc/difflex/issues"
"Documentation" = "https://github.com/kakehashi-inc/difflex#readme"

[project.optional-dependencies]
dev = [
    "pylint",
    "pylint-plugin-utils",
    "black",
    "nuitka>=2.8"
]
build = [
    "build>=1.0.0",
    "twine>=5.0.0",
    "wheel>=0.42.0"
]

[project.scripts]
difflex = "difflex.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "build>=1.0.0",
    "twine>=5.0.0",
    "wheel>=0.42.0",
]

[tool.black]
line-length = 160
exclude = 'tests/'

[tool.pylint.MASTER]
#load-plugins =

[tool.pylint.FORMAT]
max-line-length = 160

[tool.pylint.DESIGN]
max-args = 10
max-locals = 25
max-branches = 20
max-statements = 60

[tool.pylint.messages_control]
disable = [
    "W0718",
]

[tool.setuptools.packages.find]
where = ["src"]
